+++ src/main/java/com/fabianospdev/volunteer/controller/PastorController.java
@@ -1,7 +1,7 @@
 package com.fabianospdev.volunteer.controller;
 
-import com.fabianospdev.volunteer.models.Pastor;
 import com.fabianospdev.volunteer.dto.PastorDTO;
+import com.fabianospdev.volunteer.models.Pastor;
 import com.fabianospdev.volunteer.services.PastorService;
 import com.fabianospdev.volunteer.usecases.pastor.PastorUseCase;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,10 +11,11 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping(value="/pastors")
-public class PastorController {
+public class PastorController{
 
     @Autowired
     private PastorService service;
@@ -36,6 +37,11 @@ public class PastorController {
 
     @RequestMapping(value="/{id}", method=RequestMethod.GET)
     public ResponseEntity<PastorDTO> findById(@PathVariable String id) {
+
+        if(id == null || id.isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        }
+
         Pastor obj = service.findById(id);
         return ResponseEntity.ok().body(new PastorDTO(obj));
     }
@@ -48,6 +54,11 @@ public class PastorController {
 
     @RequestMapping(method=RequestMethod.POST)
     public ResponseEntity<Void> insert(@RequestBody Pastor obj) {
+
+        if(obj == null) {
+            return ResponseEntity.badRequest().build();
+        }
+
         obj = service.insert(obj);
         URI uri = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}").buildAndExpand(obj.getId()).toUri();
         return ResponseEntity.created(uri).build();
@@ -56,20 +67,47 @@ public class PastorController {
 
     @RequestMapping(value="/{id}", method=RequestMethod.DELETE)
     public ResponseEntity<Void> delete(@PathVariable String id) {
+
+        if(id == null || id.isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        }
+
         service.delete(id);
         return ResponseEntity.noContent().build();
     }
 
     @RequestMapping(value="/{id}/dto", method=RequestMethod.PUT)
     public ResponseEntity<Void> update(@RequestBody PastorDTO objDto, @PathVariable String id) {
-        Pastor obj = service.fromDTO(objDto);
-        obj.setId(id);
-        obj = service.update(obj);
-        return ResponseEntity.noContent().build();
+
+        if(objDto == null || objDto.getId() == null) {
+            return ResponseEntity.badRequest().build();
+        }
+
+        Optional<Pastor> optionalPastor = Optional.ofNullable(service.findById(id));
+
+        if(optionalPastor.isPresent()) {
+            Pastor oldObj = optionalPastor.get();
+
+            oldObj.setName(objDto.getName());
+            oldObj.setPhone(objDto.getPhone());
+            oldObj.setEmail(objDto.getEmail());
+
+            service.update(oldObj);
+
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
+
 
     @RequestMapping(value="/{id}", method=RequestMethod.PUT)
     public ResponseEntity<Void> updatefull(@RequestBody Pastor obj, @PathVariable String id) {
+
+        if(obj == null || id == null || id.isEmpty() || !id.equals(obj.getId())) {
+            return ResponseEntity.badRequest().build();
+        }
+
         obj.setId(id);
         obj = service.update(obj);
         return ResponseEntity.noContent().build();
